SSP-59 + SSP-62 Authentication login and logout with BDD
