+++ .gitignore
@@ -88,3 +88,7 @@ docs/_build/
 
 # PyBuilder
 target/
+
+# Vagrant
+*.vagrant*
+.Vagrantfile.*
