+++ README.md
@@ -1,5 +1,7 @@
 #To open plan editor run the server and open http://127.0.0.1:8000/edit_plan/
-
+1. Clone this project
+2. vagrant up
+3. vagrant ssh
 
 #REQUIREMENTS:
 * Postgres 9.3.10
