+++ SSP/plan_editor/templates/login.html
@@ -3,7 +3,7 @@
 {% block login %}
     <div class="row">
         <div class="col-md-10 col-md-offset-1">
-            <form action="{% url 'auth:login' %}" method="post" role="form" class="form-inline">
+            <form action="{% url 'login' %}" method="post" role="form" class="form-inline">
                 {% csrf_token %}
             <div class="form-group">
                 <label for="username"> Name </label>
