+++ SSP/plan_editor/urls.py
@@ -1,8 +1,7 @@
-# coding: utf-8
 from django.conf.urls import patterns, url
 
 __author__ = 'mironnn'
 
-urlpatterns = patterns('SSP.plan_editor.views',
+urlpatterns = patterns('plan_editor.views',
                        url(r'^$', 'show_map', name='show_map'),
                        )
