+++ sonar-project.properties
@@ -1,0 +1,13 @@
+# Required metadata
+sonar.projectKey=org.codehaus.sonar:python-simple-sonar-runner
+sonar.projectName=Python :: SSP Project : SonarQube Runner
+sonar.projectVersion=1.0
+
+# Comma-separated paths to directories with sources (required)
+sonar.sources=SSP
+
+# Language
+sonar.language=py
+
+# Encoding of the source files
+sonar.sourceEncoding=UTF-8
