+++ vagrant_private_key
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEogIBAAKCAQEA5wbLZEBmQtOfrH2yJ0FxsvFHnPkAAAkFY9Li2EzOjMcW/ugW
+9mnEjzygJnGuhtyr+B9lf3ZE46DTrZHyAcsH9s+qhrIexC0oJFaAEfqCZ/HOJsjd
+WcQrCqkM1g1pXO3/Z1sNBwhuVWfubgPMZxlSPV3mkdDRoz5FaqG8obj0gXr3lfir
+2MeNt8uP8kt+c7k6NPFS7mwUoFCXjqUQwwdQTHitaf2qduhlQ4euVEymDs1iIlhZ
+30ma70Xhly2X1Lhi4yhg/cCykph8U393sDWkwiSP/KKs0u9x5XewtPNVJCMN46CH
+UFXzKkmvSPUHeIIqimMW7lINfUGU3uvXvAyfvwIDAQABAoIBADCPBgsxWg0rzAEG
+fn6vy8Uv37yHvc9EFiTwO4v5QOToJ9Y2VDVA7/anaIYMGQHxuheyCbRr2Hj74eI2
+reZglXD0H0Smhb74OtrAb7TXsbdFZO5UA+qNR/q6o58OPJoy0B8aTTewNviumrf3
+BmyZLucFKKR3fsjQGukZy7s2U/Yuy0KORM0sD0eQBQppOhUHS00bzc0I+NZ8YkMz
+jaui8Xf13otiFimXnQKUTvnG+ZXS7/EcUP1zZ3y4gXHVw1nyXBYl3LjRMqC80BM4
+OVmXFGe0XflazyK9MtLOeYDwQUh0lNC1sdJqX5aeHpiOSSMDHX7PuMy2w9bPqxYI
+X4AZiCECgYEA9wAKIx0oeg1NsmoACMslBvUdJDWtKTgXPMQEO+xr3vmYyiJnJT1V
+HDUYf+Ikq65Ffw0TFwtKWTt+RDkujoUXU8MvNjL3lZu1ancrZkwFrKTylcwEEVXE
+huboZVypGzWwCIFmiomya6hevxOm6pHAe5vGsz8Ee2E3zcwMzatW75ECgYEA73HB
+uHMNSt5frFAb8uAaIbw/1/RB2VWndAltpwFzjwE6H83ne4LgV/nfwV9eRl9U6hao
+2lnaZ6r4i1pBxPtMbNYoNOhyobbDPPnQ9opVSWf7qT3Rib/OGNeLqiLdQwsNnptK
+Kh6rXzK3axzjLxvKr/Sg0dhO6siVuFLBf7IYkk8CgYApiGwr8183Br2l3oUa9W3I
+ebbou83ePHxLz3oWyW5qE/ZJJvfJH0MMdzEjb9dStv8gmgY3vcJGtuZIeoUFmgOO
+JifuKSvQkFo1CMzXarkZirLv1Vppv85IENPLF4Wn+aMcCwY/DWBfEfS8lIVgtBl1
+r3/96/7I8FP92u3cN49qwQKBgDvbRSEV6cnAg0wxDnuNmBq803zpu5hX+C9it/2a
+uEeUd/JBZ8+RbQGxHkVZ0NhotnJ6TaWf9qAZtl4npu9WHsnKU2y+ewGADZhtntYy
+yJU8IEYzhPEZeKzvHSPf6U38MmBm04hp+E8o/38QTp6VJPsNfTJYD4pTsFHOpj6E
+x+2zAoGAX5BydKUi6/BEQY7f9gLVvM0dmeQpLZZM57nbbRN6GmG/GCzlSF7qG7in
+SxDnaGovSmMmRm1XstgbUuQvuiS0TFE9OGRCUIPGp1YRN4vKG/Y1v9v4OTuOL/tn
+DrQxpIzvDnmLz+xQM0wMkhMPVHKAvLFRauWbKpP62F60jK6AWDI=
+-----END RSA PRIVATE KEY-----
